@@ -1,11 +1,9 @@
 package com.isscroberto.onemovie.movie;
 
-import android.app.Activity;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.MenuItem;
+import androidx.appcompat.widget.Toolbar;
 
 import com.github.chrisbanes.photoview.PhotoView;
 import com.isscroberto.onemovie.R;
