@@ -1,7 +1,7 @@
 package com.isscroberto.onemovie.splash;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.google.android.gms.ads.MobileAds;
